+++ simplq/src/components/pages/Admin/ShareBar.jsx
