+++ simplq/src/components/pages/Admin/index.jsx
