+++ simplq/src/components/pages/Join/index.jsx
